+++ src/server/middleware/angular-ssr.ts
@@ -1,5 +1,7 @@
 import { ApplicationRef } from '@angular/core';
 import { RequestHandler } from 'express';
+import { join } from 'path';
+import { serverConfig } from '../server-config';
 import { renderWithSsr } from '../utils/angular-ssr';
 import { pageCache } from '../utils/page-cache';
 
@@ -7,7 +9,7 @@ import { pageCache } from '../utils/page-cache';
  * Express middleware for handling Angular SSR rendering
  */
 export function angularSsrMiddleware(bootstrap: () => Promise<ApplicationRef>): RequestHandler {
-  return (req, res, next) => {
+  return (req, res) => {
     const { originalUrl, headers } = req;
     const fullUrl = `//${headers.host}${originalUrl}`;
 
@@ -21,12 +23,8 @@ export function angularSsrMiddleware(bootstrap: () => Promise<ApplicationRef>): 
     res.appendHeader('x-cache', 'MISS');
 
     // Otherwise, render the page and cache it
-    return renderWithSsr(bootstrap, fullUrl)
-      .then(html => res.send(html))
-      .catch(err => {
-        // eslint-disable-next-line no-console
-        console.error('Error during SSR:', err);
-        next(err);
-      });
+    return renderWithSsr(bootstrap, fullUrl).then(html =>
+      html ? res.send(html) : res.sendFile(join(serverConfig.browserDistFolder, 'index.csr.html')),
+    ); // send normal non SSR index.html when SSR fails.
   };
 }