+++ src/server/server-config.ts
@@ -7,6 +7,7 @@ const serverDistFolder = dirname(fileURLToPath(import.meta.url));
 export const serverConfig = {
   // Server settings
   PORT: process.env['PORT'] || 8080,
+  METRICS_PORT: process.env['METRICS_PORT'] || 9090,
   HOST: process.env['HOST'] || `localhost:${process.env['PORT'] || 8080}`,
 
   // Paths